requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cached-property==1.5.2
Django==3.2.5
django-url-filter==0.3.15
enum-compat==0.0.3
Pillow==8.3.0
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
